Refactor el módulo de verificación de URLs: corrige la importación del servicio de verificación y crea un nuevo archivo para la lógica de verificación de URLs.
